requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
pymongo==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pytest

